Fix Custom Title Rendering
Make custom back button text for iOS which just prints 'Back' so there
shouldn't be any overlap.

Conditional styling to fix apparent margin difference between the two icons

Tested on iOS and Android
