+++ react-native/navigation/addPostingsRoutes.js
@@ -1,12 +1,15 @@
 import React from 'react';
 import { View, Text, Button, TouchableOpacity, Image, StyleSheet } from 'react-native';
+import { Ionicons } from '@expo/vector-icons';
+import { HeaderBackButton } from '@react-navigation/stack';
+
 
 import PostingDetailScreen from '../screens/PostingDetailScreen';
 import EditPostingScreen from '../screens/EditPostingScreen';
 
 import Colors from '../config/colors';
 import { headerOptions }  from '../config/navigation-options';
-import IconDisplay from '../components/IconDisplay';
+import CustomHeaderTitle from '../components/CustomHeaderTitle';
 
 // Combine multiple screen routes that can be shown on different...
 // Navigation stacks.
@@ -50,19 +53,9 @@ export const addPostingsRoutes = (Stack, navigation) => {
             component={PostingDetailScreen}
             options={
                 ({route, navigation}) => ({
-                  // headerTitle: route.params.request
-                    /* ? `Request: ${route.params.title}` */
-                    /* : `Offer: ${route.params.title}`, */
-                    /* ? 'Request Detail' */
-                    /* : 'Offer Detail', */
-                  headerTitle: () =>
-                  <View style={styles.headerContainer}>
-                    <Text style={styles.titleText}>
-                      {route.params.request ? 'Request Detail' : 'Offer Detail'}
-                    </Text>
-                    <IconDisplay style={styles.icon} request={route.params.request}/>
-                  </View>,
+                  headerTitle: props => <CustomHeaderTitle request={route.params.request}/>,
                   ...headerOptions,
+                  headerBackTitle: 'Back',
                   /* headerRight: () => handleEdit(route), */
                 })
             }
@@ -85,17 +78,6 @@ export const addPostingsRoutes = (Stack, navigation) => {
 const styles = StyleSheet.create({
   headerRight: {
     paddingRight: 15,
-  },
-  headerContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  icon: {
-  },
-  titleText: {
-    color: Colors.light_shade4,
-    fontSize: 20,
-    fontWeight: 'bold',
   },
   doneButtonText: {
     color: Colors.contrast1,
